refactor: moves cms queries to central location
