+++ app/essays/[...filename]/page.tsx
@@ -1,5 +1,6 @@
 import { ClientPage } from "./client-page"
 import client from "../../../tina/__generated__/client"
+import { queryEssay } from "@/app/_queries"
 
 export async function generateStaticParams() {
   const pages = await client.queries.essayConnection()
@@ -15,9 +16,7 @@ export default async function Page({
 }: {
   params: { filename: string[] }
 }) {
-  const query = await client.queries.essay({
-    relativePath: `${params.filename}.md`,
-  })
+  const query = await queryEssay(`${params.filename}`)
 
   return <ClientPage {...query} />
 }
