+++ app/page.tsx
@@ -1,7 +1,7 @@
 import { BASE_TITLE } from "utils/constants"
 import { getPlatformList } from "utils/db-helpers"
-import client from "@/tina/__generated__/client"
 import { ClientPage } from "./client-page"
+import { queryContributors, queryHome } from "./_queries"
 
 export const metadata = {
   alternates: {
@@ -11,14 +11,10 @@ export const metadata = {
   description: "A video game database and blog website",
 }
 
-const platformList = getPlatformList()
-
 export default async function Page() {
-  const query = await client.queries.content({ relativePath: "home.md" })
-  const contribQuery = await client.queries.db_contributors({
-    relativePath: "contributors.json",
-  })
-  const contributors = contribQuery?.data?.db_contributors?.contributors
+  const query = await queryHome()
+  const contributors = await queryContributors()
+  const platformList = getPlatformList()
 
   return (
     <ClientPage
